+++ app/src/main/AndroidManifest.xml
@@ -11,11 +11,10 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.MBTI">
-        <activity android:name=".CompatibilityActivity"></activity>
         <activity android:name=".MyprofileActivity" />
-        <activity android:name=".MatchingPersonActivity" />
+        <activity android:name=".SearchingPersonActivity" />
         <activity android:name=".UpdateActivity" />
-        <activity android:name=".MatchingActivity" />
+        <activity android:name=".SearchingActivity" />
         <activity android:name=".JoinActivity" />
         <activity android:name=".MainActivity" />
         <activity android:name=".LoginActivity">
