+++ app/src/main/res/layout/activity_main.xml
@@ -41,8 +41,20 @@
 
         <LinearLayout
             android:layout_width="match_parent"
+            android:layout_height="350dp"
+            android:orientation="vertical"></LinearLayout>
+
+        <LinearLayout
+            android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:orientation="vertical"/>
+            android:orientation="vertical">
+
+            <Button
+                android:id="@+id/btn_matching"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="매칭하기" />
+        </LinearLayout>
 
     </LinearLayout>
 
@@ -52,11 +64,11 @@
         android:orientation="horizontal">
 
         <Button
-            android:id="@+id/btn_matching"
+            android:id="@+id/btn_searching"
             android:layout_width="250dp"
             android:layout_height="wrap_content"
             android:layout_margin="3dp"
-            android:text="매칭하기" />
+            android:text="검색하기" />
 
         <Button
             android:id="@+id/btn_logout2"
