+++ app/src/main/res/layout/activity_searching.xml
@@ -79,11 +79,6 @@
                     android:scaleType="fitXY"
                     app:srcCompat="@drawable/mbtichart" />
 
-                <Button
-                    android:id="@+id/btn_matchingOption2"
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="잘 맞는 유형 매칭하기" />
             </LinearLayout>
         </LinearLayout>
     </ScrollView>
